Refactor matching if commits are relevant for path filter
By filtering the git commits on paths we weren't able to find the last
version if that was filtered out. Instead loop through all commits and
implement the path matching ourselves. This should also workaround a
few bugs in the go-git implementation
