+++ .changes/unreleased/Fixed-20230317-150045.yaml
@@ -1,0 +1,3 @@
+kind: Fixed
+body: Refactor matching if commits are relevant for path filter
+time: 2023-03-17T15:00:45.506734+01:00
